rename calculation_preparation.hpp to calculation_input_preparation.hpp
